requirements.txt
# Django & Django REST Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Static file handling
whitenoise==6.5.0
django-storages==1.14.2
boto3==1.28.62

# GIS
django-leaflet>=0.28.0
# GDAL==3.7.2  # Install via homebrew first: brew install gdal
pyproj==3.6.0

# Music API integrations
spotipy==2.23.0
# applemusicpy==0.3.3  # Comment out for now
# soundcloud-api==0.3.8  # Comment out for now

# Authentication & Security
django-allauth==0.57.0
# django-rest-auth==0.9.5  # Deprecated package, using django-allauth instead
oauthlib==3.2.2
django-oauth-toolkit==2.3.0
social-auth-app-django==5.3.0

# Environment variables
python-dotenv==1.0.0
python-decouple==3.8

# Documentation
drf-spectacular==0.26.4
drf-yasg==1.21.7
Markdown==3.5

# Image processing
Pillow==10.0.1

# Push notifications
firebase-admin==6.2.0
# pyfcm==1.5.4  # Comment out for now

# Async tasks & scheduled jobs
celery==5.3.4
django-celery-beat==2.5.0
redis==5.0.1

# Websockets for real-time features
channels>=4.0.0
channels-redis>=4.1.0
daphne==4.0.0

# Performance & Caching
django-redis==5.4.0
django-cachalot==2.6.1

# Development tools
django-debug-toolbar==4.2.0
black==23.9.1
isort==5.12.0
flake8==6.1.0

# Testing
pytest==7.4.2
pytest-django==4.5.2
coverage==7.3.2
factory-boy==3.3.0

# Production
gunicorn==21.2.0
uvicorn==0.23.2

# GIS packages
djangorestframework-gis>=1.0.0 